requirements.txt
Django
south
nodeenv
django-mptt
Pillow
